Add a 'Removed' DeviceType
The X.org server sends HierarchyChanged events with device type set to
zero for devices that were removed. Since "DeviceType" is used in an
enum, this means that only values of that enum can appear. This is not
allowed by the spec, but it is what happens in reality. Thus, this
commit changes xinput.xml to describe reality by adding "Removed" as a
new "DeviceType".

Originally-reported: https://github.com/psychon/x11rb/issues/458
Signed-off-by: Uli Schlachter <psychon@znc.in>