xkb: Make all the breaking changes
An update to xkb.xml is long overdue, but has been put on hold to avoid
breaking existing users of the library.

It has been almost a decade since the comment "This pad is a workaround to
prevent an API break" was added. One API adjustment per decade sounds about
right, so here we go:

1. Use <pad align> instead of a complex alignment_pad calculation that
   predates the existence of <pad align> in CountedString16.
2. Uncomment Property now that CountedString16 works as expected.
3. Rewrite Doodad to use <switch> instead of <union>. This is necessary
   because the various types of Doodad are not all the same length on the
   wire (eg. CountedString16 can be an arbitrary length). And, in general,
   union is useless for pretty-printers like Wireshark and languages other
   than C.
4. Uncomment modLatches in LatchLockState
5. Uncomment GetGeometry and SetGeometry requests

Note that this is a breaking change. The libxcb maintainers will need to
bump the API and ABI version numbers.